Add a unit test for unsupported mixed interleave modes (#201)
CharLS doesn't support mixing interleave modes.
Add a unit test to ensure that CharLS will throw a not supported exception.
The actual checking was already done.

Note: also resolve some static analyzer warnings, reported with the latest version of ReSharper\clang-tidy.

+++ benchmark/context_regular_mode_v220.h
@@ -64,7 +64,7 @@ struct jls_context_v220 final
             {
                 b = -n + 1;
             }
-            C = C - static_cast<int16_t>(C > -128);
+            C = static_cast<int16_t>(C - static_cast<int16_t>(C > -128));
         }
         else if (b > 0)
         {
@@ -73,7 +73,7 @@ struct jls_context_v220 final
             {
                 b = 0;
             }
-            C = C + static_cast<int16_t>(C < 127);
+            C = static_cast<int16_t>(C + static_cast<int16_t>(C < 127));
         }
         B = b;
